code refactoring 2
